Unified code with GraphQLParser
