+++ layers/Engine/packages/graphql-parser/src/ExtendedSpec/Constants/QuerySyntax.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace PoP\ComponentModel\Constants;
+namespace PoP\GraphQLParser\Constants;
 
 class QuerySyntax
 {
